Fix and organize testapp
