+++ src/testapp/sessions/vehicles.cpp
@@ -22,9 +22,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-#include "scene_vehicles.h"
-#include "scenarios.h"
-#include "identifiers.h"
+#include "vehicles.h"
 
 #include <adera/activescene/VehicleBuilder.h>
 #include <adera/machines/links.h>
