+++ src/testapp/testapp.cpp
@@ -23,8 +23,7 @@
  * SOFTWARE.
  */
 #include "testapp.h"
-
-#include "activescenes/identifiers.h"
+#include "identifiers.h"
 
 #include <osp/core/Resources.h>
 #include <osp/drawing/own_restypes.h>
